calculate exact buffer size using inputs
previously we had a default buffer size (arbitrarily set to 1024 bytes).
this was a known placeholder and the time finally came to replace it!

i ended up with a little procedure for calculating the exact buffer
size using the different kinds of inputs we store in the
resulting buffer.

since we operate on a few recurring types, i decided to represent those
in the procedure:
* integers (stored as varints, shorthand `v`),
* bytes (stored as they are, shorthand `b`),
* lists of HASH_SIZE buffers (shorthand `h`),
* and strings (converted into buffers, with a stringLen varint
  prepended, shorthand `s`).

the routine, determineBufferSize, takes these four different types and
uses them to calculate the correct buffer size. input looks like the
following - a list of objects, with type information included:
const size = determineBufferSize([
    {v: constants.TOPIC_POST},
    {h: links.length}
    {s: channelBuf.length},
    {s: topicBuf.length}
])

SHORTHAND:
v: integer which will be encoded as a varint
b: number of bytes
s: string input.
   rep.s contains the length of the corresponding string buffer
   e.g. rep.s === channelBuf.length
h: represents an array of HASH_SIZE sized hashes.
   rep.h is the number of such entries in the actual array
   e.g. rep.h === hashes.length